+++ src/components/WorkProfile_FormController/FormController.scss
@@ -1,0 +1,34 @@
+
+.indicators-container {
+    position: absolute;
+    top: 0;
+    opacity: 40%;
+    height: fit-content;
+    transition: all 0.3s ease-in-out;
+  
+    .form-indicator {
+      background-color: #b3e5be;
+      min-height: 8px;
+      width: 90%;
+      &:hover {
+        opacity: 100%;
+        transform: scale(1.05);
+      }
+    }
+  
+    .active {
+      background-color: #2738f5;
+    }
+  
+    &:hover {
+      opacity: 100%;
+      transform: scale(1.05);
+    }
+  }
+  .carousel-item {
+    padding-top: 50px;
+  }
+
+  .carousel-item {
+    padding-top: 50px;
+  }
