+++ src/views/home/home.js
@@ -6,14 +6,18 @@ import SectionHomeSystem from "../../components/Section_home_system/index.js";
 import Banner from "../../components/Banner/Banner.js";
 import BigBanner from "../../components/BigBanner/BigBanner.js";
 import ContactForm from "../../components/ContactForm/index.js";
-import ModalForm from "../../components/modalForm/ModalForm";
+import LayoutWorkProfile from "../../components/WorkProfile_Layout/LayoutWorkProfile";
 
 // encargada de renderizar todos los componentes de la vista home
 const Ui_home = () => {
+
+  const handleOpen = (ModalOpen) => {
+    console.log('modal abierto?? ' + ModalOpen)
+  };
+
   return (
     <div>
-      <ModalForm/>
-      {/* <Navbar />
+      <Navbar />
       <BigBanner />
       <Banner />
       <SectionSinJrsNoHaySrs />
@@ -25,7 +29,7 @@ const Ui_home = () => {
           <ContactForm />
         </div>
       </div>
-      <Footer /> */}
+      <Footer />
     </div>
   );
 };
